* Modified multicast and broadcast transports to convert to big endian * Modified all of MADARA to use stdint.h integers where appropriate, especially when it comes to what might be transported * Updated all MADARA functions to use Madara::Knowledge_Record::VALUE_TYPE where appropriate, in case we want to implement floating point or other type knowledge bases * Updated MADARA version to 0.8.4
git-svn-id: http://madara.googlecode.com/svn/trunk@561 2850746a-f1c8-11dd-b507-8b4ecd5210be

+++ include/madara/Globals.h
@@ -13,6 +13,7 @@
 #include <string>
 #include <vector>
 #include <map>
+#include <stdint.h>
 
 namespace Madara
 {
@@ -44,7 +45,7 @@ namespace Madara
         quality (0), write_quality (0) 
     {}
 
-    typedef  long long      VALUE_TYPE;
+    typedef  int64_t      VALUE_TYPE;
     
     static bool lt (VALUE_TYPE lhs, VALUE_TYPE rhs);
 
@@ -69,22 +70,22 @@ namespace Madara
     /**
      * last modification time
      **/
-    unsigned long long clock;
+    uint64_t clock;
 
     /**
      * scope (global or local)
      **/
-    int scope;
+    int32_t scope;
 
     /**
      * priority of the update
      **/
-    unsigned long quality;
+    uint32_t quality;
 
     /**
      * write priority for any local updates
      **/
-    unsigned long write_quality;
+    uint32_t write_quality;
   };
 
   typedef ::std::map < ::std::string, Knowledge_Record> Knowledge_Map;
